rename sequences
