+++ src/mrseq/sequences/t2star_multi_echo_flash.py
