+++ tests/sequences/test_radial_flash.py
@@ -1,9 +1,10 @@
-"""Tests for Gold standard SE-based T1rho mapping sequence."""
+"""Tests for radial FLASH sequence."""
 
+import numpy as np
 import pytest
-from mrseq.scripts.t1rho_se_single_line import main as create_seq
+from mrseq.sequences.radial_flash import main as create_seq
 
-EXPECTED_DUR = 3072.000000  # defined 2025-02-20
+EXPECTED_DUR = 0.75043  # defined 2025-10-17
 
 
 def test_default_seq_duration(system_defaults):
@@ -22,20 +23,31 @@ def test_seq_creation_error_on_short_te(system_defaults):
 def test_seq_creation_error_on_short_tr(system_defaults):
     """Test if error is raised on too short repetition time."""
     with pytest.raises(ValueError):
-        create_seq(system=system_defaults, tr=5e-3, show_plots=False)
+        create_seq(system=system_defaults, tr=2e-3, show_plots=False)
 
 
-def test_seq_duration_vary_params_without_effect(system_defaults):
-    """Test if sequence duration is as expected."""
+def test_seq_predefined_echo_time(system_defaults):
+    """Test sequence with predefined echo time."""
     seq, _ = create_seq(
         system=system_defaults,
-        te=10e-3,  # default None
-        fov_xy=192e-3,  # default 128e-3
-        n_readout=192,  # default 128
-        slice_thickness=6e-3,  # default 8e-3
+        te=3e-3,
+        show_plots=False,
+        test_report=False,
+        timing_check=False,
+    )
+    assert seq
+
+
+def test_seq_m2d(system_defaults):
+    """Test multi-slice acquisition."""
+    n_slices = 8
+    seq, _ = create_seq(
+        system=system_defaults,
+        n_slices=n_slices,
         show_plots=False,
         test_report=False,
         timing_check=False,
     )
     duration = seq.duration()[0]
-    assert duration == pytest.approx(EXPECTED_DUR)
+    # we need a larger tolerance because noise samples are only acquired once and not for each slice
+    np.testing.assert_allclose(duration / n_slices, EXPECTED_DUR, rtol=0.01)
