+++ tests/sequences/test_t1_molli_bssfp.py
@@ -1,9 +1,9 @@
 """Tests for 2D Cartesian FLASH with T2-preparation pulses for T2 mapping."""
 
 import pytest
-from mrseq.scripts.t2star_multi_echo_flash import main as create_seq
+from mrseq.sequences.t1_molli_bssfp import main as create_seq
 
-EXPECTED_DUR = 2.83989  # defined 2025-11-22
+EXPECTED_DUR = 10.95339  # defined 2025-11-24
 
 
 def test_default_seq_duration(system_defaults):
@@ -23,11 +23,3 @@ def test_seq_creation_error_on_short_tr(system_defaults):
     """Test if error is raised on too short repetition time."""
     with pytest.raises(ValueError):
         create_seq(system=system_defaults, tr=2e-3, show_plots=False)
-
-
-def test_seq_creation_error_on_wrong_partial_echo_factor(system_defaults):
-    """Test if error is raised on wrong partial echo factor."""
-    with pytest.raises(ValueError):
-        create_seq(system=system_defaults, partial_echo_factor=1.1, show_plots=False)
-    with pytest.raises(ValueError):
-        create_seq(system=system_defaults, partial_echo_factor=0.4, show_plots=False)
