+++ tests/sequences/test_t2_t2prep_flash.py
@@ -1,10 +1,10 @@
-"""Tests for radial FLASH sequence."""
+"""Tests for 2D Cartesian FLASH with T2-preparation pulses for T2 mapping."""
 
 import numpy as np
 import pytest
-from mrseq.scripts.radial_flash import main as create_seq
+from mrseq.sequences.t2_t2prep_flash import main as create_seq
 
-EXPECTED_DUR = 0.75043  # defined 2025-10-17
+EXPECTED_DUR = 8.17667  # defined 2026-01-26
 
 
 def test_default_seq_duration(system_defaults):
@@ -26,28 +26,14 @@ def test_seq_creation_error_on_short_tr(system_defaults):
         create_seq(system=system_defaults, tr=2e-3, show_plots=False)
 
 
-def test_seq_predefined_echo_time(system_defaults):
-    """Test sequence with predefined echo time."""
+def test_seq_duration_vary_params_without_effect(system_defaults):
+    """Test if sequence duration is as expected."""
     seq, _ = create_seq(
         system=system_defaults,
-        te=3e-3,
-        show_plots=False,
-        test_report=False,
-        timing_check=False,
-    )
-    assert seq
-
-
-def test_seq_m2d(system_defaults):
-    """Test multi-slice acquisition."""
-    n_slices = 8
-    seq, _ = create_seq(
-        system=system_defaults,
-        n_slices=n_slices,
+        t2_prep_echo_times=np.array([0.05, 0.1, 0.2]),
         show_plots=False,
         test_report=False,
         timing_check=False,
     )
     duration = seq.duration()[0]
-    # we need a larger tolerance because noise samples are only acquired once and not for each slice
-    np.testing.assert_allclose(duration / n_slices, EXPECTED_DUR, rtol=0.01)
+    assert duration == pytest.approx(EXPECTED_DUR)
